backend/pyproject.toml
[project]
name = "agentic-crawler"
version = "1.0.0"
description = "Autonomous evaluation agent for comprehensive website testing and analysis"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["qa", "testing", "web-crawler", "automation", "ai-agent"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "firecrawl-py>=4.5.0",
    "ipykernel>=7.1.0",
    "langchain>=1.0.2",
    "langchain-google-genai>=3.0.0",
    "langchain-groq>=1.0.0",
    "langchain-mcp-adapters>=0.1.11",
    "mcp>=1.19.0",
    "rich>=14.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "deepagents>=0.2.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
agentic-crawler = "agentic_crawler.cli:main"
agentic-crawler-web = "agentic_crawler.web.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/agentic-crawler"
Documentation = "https://github.com/yourusername/agentic-crawler#readme"
Repository = "https://github.com/yourusername/agentic-crawler"
Issues = "https://github.com/yourusername/agentic-crawler/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_crawler"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false


